Finished cleaning up UI and fixed edit plant address not updating
There was a bug when Editing Plant that set the lat/long to zero when updating it. Also the production info was missing from the view plants screen and subsequently the edit plant screen which caused you to have to reenter the information every time you edited. Fixed this issue as well.